split read/write tests
