+++ integration_tests/src/main/python/orc_write_test.py
@@ -1,0 +1,81 @@
+# Copyright (c) 2020, NVIDIA CORPORATION.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+import pytest
+
+from asserts import assert_gpu_and_cpu_writes_are_equal_collect
+from datetime import date, datetime, timezone
+from data_gen import *
+from marks import *
+from pyspark.sql.types import *
+
+orc_write_gens_list = [
+        [byte_gen, short_gen, int_gen, long_gen, float_gen, double_gen,
+            string_gen, boolean_gen, DateGen(start=date(1590, 1, 1)),
+            TimestampGen(start=datetime(1970, 1, 1, tzinfo=timezone.utc))],
+        pytest.param([date_gen], marks=pytest.mark.xfail(reason='https://github.com/NVIDIA/spark-rapids/issues/139')),
+        pytest.param([timestamp_gen], marks=pytest.mark.xfail(reason='https://github.com/NVIDIA/spark-rapids/issues/140'))]
+
+@pytest.mark.parametrize('orc_gens', orc_write_gens_list, ids=idfn)
+def test_write_round_trip(spark_tmp_path, orc_gens):
+    gen_list = [('_c' + str(i), gen) for i, gen in enumerate(orc_gens)]
+    data_path = spark_tmp_path + '/ORC_DATA'
+    assert_gpu_and_cpu_writes_are_equal_collect(
+            lambda spark, path: gen_df(spark, gen_list).coalesce(1).write.orc(path),
+            lambda spark, path: spark.read.orc(path),
+            data_path)
+
+orc_part_write_gens = [
+        byte_gen, short_gen, int_gen, long_gen, float_gen, double_gen, boolean_gen,
+        # Some file systems have issues with UTF8 strings so to help the test pass even there
+        StringGen('(\\w| ){0,50}'),
+        # Once https://github.com/NVIDIA/spark-rapids/issues/139 is fixed replace this with
+        # date_gen
+        DateGen(start=date(1590, 1, 1)),
+        # Once https://github.com/NVIDIA/spark-rapids/issues/140 is fixed replace this with
+        # timestamp_gen 
+        TimestampGen(start=datetime(1970, 1, 1, tzinfo=timezone.utc))]
+
+# There are race conditions around when individual files are read in for partitioned data
+@ignore_order
+@pytest.mark.parametrize('orc_gen', orc_part_write_gens, ids=idfn)
+def test_part_write_round_trip(spark_tmp_path, orc_gen):
+    gen_list = [('a', RepeatSeqGen(orc_gen, 10)),
+            ('b', orc_gen)]
+    data_path = spark_tmp_path + '/ORC_DATA'
+    assert_gpu_and_cpu_writes_are_equal_collect(
+            lambda spark, path: gen_df(spark, gen_list).coalesce(1).write.partitionBy('a').orc(path),
+            lambda spark, path: spark.read.orc(path),
+            data_path)
+
+orc_write_compress_options = ['none', 'uncompressed', 'snappy']
+@pytest.mark.parametrize('compress', orc_write_compress_options)
+def test_compress_write_round_trip(spark_tmp_path, compress):
+    data_path = spark_tmp_path + '/ORC_DATA'
+    assert_gpu_and_cpu_writes_are_equal_collect(
+            lambda spark, path : binary_op_df(spark, long_gen).coalesce(1).write.orc(path),
+            lambda spark, path : spark.read.orc(path),
+            data_path,
+            conf={'spark.sql.orc.compression.codec': compress})
+
+@pytest.mark.parametrize('orc_gens', orc_write_gens_list, ids=idfn)
+def test_write_save_table(spark_tmp_path, orc_gens, spark_tmp_table_factory):
+    gen_list = [('_c' + str(i), gen) for i, gen in enumerate(orc_gens)]
+    data_path = spark_tmp_path + '/ORC_DATA'
+    all_confs={'spark.sql.sources.useV1SourceList': "orc"}
+    assert_gpu_and_cpu_writes_are_equal_collect(
+            lambda spark, path: gen_df(spark, gen_list).coalesce(1).write.format("orc").mode('overwrite').option("path", path).saveAsTable(spark_tmp_table_factory.get()),
+            lambda spark, path: spark.read.orc(path),
+            data_path,
+            conf=all_confs)
